requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.12
# pickle
pyaml
dill
# sklearn
mlflow
dagshub
fastapi
uvicorn


#this triggers the setup.py script. I comment it out for now
##-e .